Move common camera tools out of rtsp agent
